split god file orchestrator.py
