+++ backend/database/queries/conversations.py
@@ -169,7 +169,7 @@ async def set_workflow_state(conv_id: str, workflow_id: str, payload: dict | Non
     Caller must hold ``backend.core.locks.workflow_state_lock(conv_id, workflow_id)``
     across the read-then-write the payload was computed from. Acquisition
     sites: ``backend.api.routes.workflows.api_trigger_workflow`` and the pre/post pipeline
-    hook loops in ``backend.pipeline.orchestrator``. Direct use outside those paths
+    hook loops in ``backend.pipeline.workflow_bridge``. Direct use outside those paths
     re-introduces the read-modify-write clobber.
     """
     async with get_db() as db:
