+++ backend/pipeline/context.py
@@ -1,0 +1,365 @@
+"""
+context.py — Inbound turn assembly: load → prefixes → pre-pipeline setup.
+
+Two phases, kept in one module because both are "everything before the passes
+run":
+
+* **Load** — :func:`_load_pipeline_context` fetches every per-conversation input
+  (settings, conversation, card, director state, fragments, phrase bank,
+  lorebook, the writer/agent LLM clients) into the frozen
+  :class:`PipelineContext`, and :func:`_build_prefixes` builds the writer +
+  optional agent message prefixes from it.
+* **Prepare** — :func:`_prepare_turn` runs the pre-pipeline workflow hooks (which
+  may stream SSE events and extend the tool map / system prompt), computes the
+  lorebook block (or the agentic catalog), builds the byte-identical tool blob,
+  and yields a single :class:`_TurnSetup`.
+
+``LLMClient`` is constructed here (and *only* here) — tests patch
+``backend.pipeline.context.LLMClient`` to substitute the streaming client.
+"""
+
+from __future__ import annotations
+
+from dataclasses import dataclass
+from types import MappingProxyType
+from typing import Any, AsyncIterator, Mapping, Optional, Sequence
+
+from .. import database as db
+from ..core import ChatMessage, Macros
+from ..database.models import (
+    ActiveLorebookEntryRow,
+    CharacterCardRow,
+    ConversationRow,
+    DirectorStateRow,
+    InteractiveFragmentRow,
+    MoodFragmentRow,
+    PhraseGroup,
+    SettingsRow,
+    UserPersonaRow,
+)
+from ..inference import (
+    AbortToken,
+    LLMClient,
+    _KVCacheTracker,
+    build_prefix,
+    compute_lorebook_injection_block,
+)
+from .config import _build_writer_tools_blob
+from .passes.director import _agentic_lorebook_active, build_lorebook_catalog
+from .predicates import agent_enabled, resolve_persona_id
+from .workflow_bridge import _iterate_pre_pipeline_hooks
+
+
+@dataclass(frozen=True)
+class PipelineContext:
+    """Resolved per-conversation inputs the pipeline needs, loaded once by
+    :func:`_load_pipeline_context` and threaded through every entry point.
+
+    Frozen so the *binding* of each field is immutable; the optional fields make
+    explicit what was previously only implied by the ``ctx[...]`` vs
+    ``ctx.get(...)`` split in readers. ``card`` / ``active_persona`` are None when
+    the conversation has no character card / no active user persona; ``agent_client``
+    and ``agent_system_prompt`` are None unless a separate agent endpoint is
+    configured (they travel together). ``director`` is a mutable dict held by
+    reference and deliberately mutated in place — the regenerate paths reset its
+    ``active_moods`` / ``progressive_fields`` to the branch baseline before a turn,
+    which the frozen dataclass does not prevent (it guards rebinding the field,
+    not mutating the dict it points at).
+    """
+
+    settings: SettingsRow
+    conv: ConversationRow
+    card: Optional[CharacterCardRow]
+    director: DirectorStateRow
+    mood_fragments: list[MoodFragmentRow]
+    interactive_fragments: list[InteractiveFragmentRow]
+    phrase_bank: list[PhraseGroup]
+    lorebook_entries: list[ActiveLorebookEntryRow]
+    client: LLMClient
+    system_prompt: str
+    char_persona: str
+    mes_example: str
+    active_persona: Optional[UserPersonaRow]
+    agent_client: Optional[LLMClient]
+    agent_system_prompt: Optional[str]
+
+
+async def _load_pipeline_context(conversation_id: str, *, abort_token: AbortToken | None = None) -> PipelineContext | None:
+    """Load all per-conversation inputs needed to run the pipeline.
+
+    Fetches settings, conversation, character card, director state, fragments,
+    phrase bank, lorebook entries, and builds LLM clients. Both the writer and
+    agent clients share the same *abort_token* so a single ``/stop`` cancels
+    every pass. Creates a private token when none is supplied.
+
+    Returns a :class:`PipelineContext`, or ``None`` if the conversation was not found.
+
+    Called by every public entry point before any pipeline work begins.
+    """
+    abort_token = abort_token or AbortToken()
+    settings = await db.get_settings()
+    conv = await db.get_conversation(conversation_id)
+    if not conv:
+        return None
+
+    director = await db.get_director_state(conversation_id)
+    mood_fragments = await db.get_mood_fragments()
+    mood_fragments = [f for f in mood_fragments if f.get("enabled", True)]
+    # Prune active moods that reference disabled fragments.
+    if director and director.get("active_moods"):
+        enabled_ids = {f["id"] for f in mood_fragments}
+        director["active_moods"] = [mood for mood in director["active_moods"] if mood in enabled_ids]
+    interactive_fragments = await db.get_interactive_fragments()
+    interactive_fragments = [df for df in interactive_fragments if df.get("enabled", True)]
+    phrase_bank = await db.get_phrase_bank()
+    lorebook_entries = await db.get_active_lorebook_entries()
+    client = LLMClient(
+        settings["endpoint_url"],
+        api_key=settings.get("api_key", ""),
+        abort_token=abort_token,
+    )
+
+    card_id = conv.get("character_card_id")
+    card = await db.get_character_card(card_id) if card_id else None
+    system_prompt, char_persona, mes_example = await db.resolve_char_context(conv, settings, card=card)
+
+    active_persona = None
+    active_persona_id = resolve_persona_id(conv, card, settings)
+    if active_persona_id:
+        active_persona = await db.get_user_persona(active_persona_id)
+
+    agent_same = settings.get("agent_same_as_writer", True)
+    agent_client = None
+    agent_system_prompt = None
+    if not agent_same and settings.get("agent_endpoint_id"):
+        agent_url = settings.get("agent_endpoint_url", settings["endpoint_url"])
+        agent_api_key = settings.get("agent_api_key", settings.get("api_key", ""))
+        agent_client = LLMClient(
+            agent_url,
+            api_key=agent_api_key,
+            abort_token=abort_token,
+        )
+        agent_system_prompt, _, _ = await db.resolve_char_context(
+            conv, settings, shared_key="agent_shared_system_prompt", card=card
+        )
+
+    return PipelineContext(
+        settings=settings,
+        conv=conv,
+        card=card,
+        director=director,
+        mood_fragments=mood_fragments,
+        interactive_fragments=interactive_fragments,
+        phrase_bank=phrase_bank,
+        lorebook_entries=lorebook_entries,
+        client=client,
+        system_prompt=system_prompt,
+        char_persona=char_persona,
+        mes_example=mes_example,
+        active_persona=active_persona,
+        agent_client=agent_client,
+        agent_system_prompt=agent_system_prompt,
+    )
+
+
+def _build_prefix_from_ctx(
+    ctx: PipelineContext,
+    history: Sequence[Mapping[str, Any]],
+    *,
+    system_prompt: str | None = None,
+    extra_system_blocks: list[str] | None = None,
+) -> list[ChatMessage]:
+    """Build the LLM message prefix (system prompt + chat history) from *ctx*.
+
+    *system_prompt* overrides ``ctx.system_prompt`` when provided — used for
+    the agent prefix in dual-model mode. *extra_system_blocks* appends
+    additional system sections contributed by pre-pipeline hooks.
+
+    Called by :func:`_build_prefixes`.
+    """
+    conv = ctx.conv
+    active_persona = ctx.active_persona
+    macros = Macros.from_settings(ctx.settings, conv["character_name"], active_persona)
+    user_description = active_persona.get("description", "") if active_persona else ctx.settings.get("user_description", "")
+
+    return build_prefix(
+        system_prompt if system_prompt is not None else ctx.system_prompt,
+        ctx.char_persona,
+        conv["character_scenario"],
+        ctx.mes_example,
+        ("" if ctx.settings.get("prevent_prompt_overrides") else conv.get("post_history_instructions", "")),
+        history,
+        macros,
+        user_description,
+        extra_system_blocks=extra_system_blocks,
+    )
+
+
+def _build_prefixes(
+    ctx: PipelineContext,
+    history: Sequence[Mapping[str, Any]],
+    *,
+    extra_system_blocks: list[str] | None = None,
+) -> tuple[list[ChatMessage], list[ChatMessage] | None]:
+    """Build the writer prefix and optional agent prefix for a turn.
+
+    Returns ``(prefix, agent_prefix)``. ``agent_prefix`` is ``None`` when no
+    separate agent system prompt is configured (single-model mode).
+    *extra_system_blocks* from pre-pipeline hooks is applied to both so the
+    system body stays identical across all passes.
+
+    Called by :func:`_prepare_turn`.
+    """
+    prefix = _build_prefix_from_ctx(ctx, history, extra_system_blocks=extra_system_blocks)
+    agent_sp = ctx.agent_system_prompt
+    agent_prefix = (
+        _build_prefix_from_ctx(
+            ctx,
+            history,
+            system_prompt=agent_sp,
+            extra_system_blocks=extra_system_blocks,
+        )
+        if agent_sp is not None
+        else None
+    )
+    return prefix, agent_prefix
+
+
+def _compute_lorebook(macros: Macros, ctx: PipelineContext, messages: Sequence[Mapping[str, Any]]) -> str:
+    """Scan *messages* for lorebook keyword matches and return the injection block.
+
+    Called by :func:`_prepare_turn` when agentic lorebook mode is off.
+    """
+    return compute_lorebook_injection_block(
+        messages,
+        ctx.lorebook_entries,
+        macros,
+    )
+
+
+@dataclass
+class _TurnSetup:
+    """Resolved per-turn pipeline inputs produced by :func:`_prepare_turn`.
+
+    Bundles everything the entry points compute identically between persisting
+    the user row and launching ``_run_pipeline``: the (writer, agent) prefixes
+    with any pre-pipeline ``system_prompt`` blocks already applied, the merged
+    tool-enable map, and the per-turn shared identities (macros, lorebook
+    block, scratch dict, KV tracker, dynamic-schema map).
+    """
+
+    prefix: list[ChatMessage]
+    agent_prefix: list[ChatMessage] | None
+    merged_enabled_tools: dict[str, bool]
+    macros: Macros
+    lorebook_block: str
+    turn_scratch: dict
+    kv_tracker: _KVCacheTracker
+    schema_overrides: Mapping[str, dict]
+    # Agentic-lorebook activation: when active, ``lorebook_block`` is "" (the
+    # keyword scan is bypassed), ``lorebook_catalog`` carries the Director's
+    # candidate catalog, and the writer block is computed post-director from the
+    # selection. When inactive both are inert (catalog "", flag False).
+    lorebook_catalog: str = ""
+    agentic_lorebook_active: bool = False
+
+
+async def _prepare_turn(
+    ctx: PipelineContext,
+    conversation_id: str,
+    *,
+    history: Sequence[Mapping[str, Any]],
+    settings: Mapping[str, Any],
+    last_user_message: str,
+    lorebook_messages: Sequence[Mapping[str, Any]],
+) -> AsyncIterator[dict | _TurnSetup]:
+    """Prepare everything a turn needs before the pipeline starts.
+
+    Builds macros, prefixes, tool maps, the lorebook block, runs pre-pipeline
+    workflow hooks (which may stream SSE events), then yields a single
+    :class:`_TurnSetup` as the last item.
+
+    Drain it as::
+
+        setup = None
+        async for ev in _prepare_turn(...):
+            if isinstance(ev, _TurnSetup):
+                setup = ev
+            else:
+                yield ev
+        assert setup is not None
+
+    Called by ``entrypoints._generate_reply`` for every generating entry point.
+    """
+    macros = Macros.from_settings(ctx.settings, ctx.conv["character_name"], ctx.active_persona)
+
+    prefix_base, agent_prefix_base = _build_prefixes(ctx, history)
+
+    turn_scratch: dict = {}
+    kv_tracker = _KVCacheTracker(conversation_id=conversation_id)
+    # Built once; when the agent is off, all tools are force-disabled.
+    enabled_tools_setting = settings.get("enabled_tools") or {}
+    if agent_enabled(settings):
+        enabled_tools_pre_merge = dict(enabled_tools_setting)
+    else:
+        enabled_tools_pre_merge = {k: False for k in enabled_tools_setting}
+
+    # When agentic lorebook is active the keyword scan is skipped; the Director
+    # picks entries from a catalog instead and the writer block is built post-director.
+    agentic_active = _agentic_lorebook_active(
+        settings, enabled_tools_pre_merge, ctx.lorebook_entries, agent_on=agent_enabled(settings)
+    )
+    if agentic_active:
+        lorebook_block = ""
+        lorebook_catalog = build_lorebook_catalog(ctx.lorebook_entries)
+    else:
+        lorebook_block = _compute_lorebook(macros, ctx, lorebook_messages)
+        lorebook_catalog = ""
+
+    # Builds direct_scene + optionally give_feedback; must be called once so all
+    # passes get byte-identical tool blobs (KV cache Invariants 3 & 5).
+    overrides = _build_writer_tools_blob(
+        settings, ctx.interactive_fragments, enabled_tools_pre_merge, agentic_lorebook=agentic_active
+    )
+    schema_overrides = MappingProxyType(overrides)
+    accumulators = {
+        "merged_enabled_tools": dict(enabled_tools_pre_merge),
+        "extras": [],
+    }
+
+    # Pre-pipeline hooks may extend the tool map or append system blocks.
+    async for ev in _iterate_pre_pipeline_hooks(
+        conversation_id=conversation_id,
+        character_id=ctx.conv.get("character_card_id"),
+        card=ctx.card,
+        history=history,
+        last_user_message=last_user_message,
+        settings=settings,
+        prefix_base=prefix_base,
+        enabled_tools_pre_merge=enabled_tools_pre_merge,
+        turn_scratch=turn_scratch,
+        client=ctx.client,
+        kv_tracker=kv_tracker,
+        schema_overrides=schema_overrides,
+        accumulators=accumulators,
+    ):
+        yield ev
+
+    extras = accumulators["extras"]
+    if extras:
+        prefix, agent_prefix = _build_prefixes(ctx, history, extra_system_blocks=extras)
+    else:
+        prefix, agent_prefix = prefix_base, agent_prefix_base
+
+    yield _TurnSetup(
+        prefix=prefix,
+        agent_prefix=agent_prefix,
+        merged_enabled_tools=accumulators["merged_enabled_tools"],
+        macros=macros,
+        lorebook_block=lorebook_block,
+        turn_scratch=turn_scratch,
+        kv_tracker=kv_tracker,
+        schema_overrides=schema_overrides,
+        lorebook_catalog=lorebook_catalog,
+        agentic_lorebook_active=agentic_active,
+    )
