+++ tests/integration/test_continue_log_turn_index.py
@@ -13,7 +13,7 @@ fresh-turn path.
 from __future__ import annotations
 
 import backend.database as dbmod
-from backend.pipeline.orchestrator import handle_turn
+from backend.pipeline import handle_turn
 
 
 async def _drain(agen) -> list[dict]:
