+++ tests/integration/workflows/test_workflow_character_fields.py
@@ -20,7 +20,8 @@ from backend.database import (
     set_active_leaf,
 )
 from backend.inference import LLMClient, _KVCacheTracker
-from backend.pipeline.orchestrator import _iterate_pre_pipeline_hooks, _run_pipeline
+from backend.pipeline.orchestrator import _run_pipeline
+from backend.pipeline.workflow_bridge import _iterate_pre_pipeline_hooks
 
 from ._fixtures import make_workflow, register_for_test
